feat: Add support for keyboard input
This commit adds a feature that allows users to input values using their
keyboard in addition to clicking the calculator buttons. The allowedKeys
array defines the keys that are allowed for input. If the user presses a
key that is not in the allowedKeys array, the input is ignored.

When the user presses a valid key, the corresponding calculator button is
clicked programmatically. For example, pressing the "Enter" key performs
a calculation, pressing the "Backspace" key deletes the previous input,
and pressing the "Delete" key clears the entire input field.

This feature improves the usability and accessibility of the calculator,
making it easier for users to perform calculations.